requirements.txt
telethon==1.36.0
python-telegram-bot==21.7
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
google-generativeai==0.8.3
apscheduler==3.10.4
pyyaml==6.0.2
python-dotenv==1.0.1
httpx==0.27.0
playwright==1.49.0
